Struct DID_NEAR_rust::DID

pub struct DID {
    pub status: UnorderedMap<String, Status>,
    pub contexts: UnorderedMap<String, Vec<String>>,
    pub public_key: UnorderedMap<String, PublicKeyList>,
    pub authentication: UnorderedMap<String, Vec<u32>>,
    pub controller: UnorderedMap<String, Vec<String>>,
    pub service: UnorderedMap<String, Vec<Service>>,
    pub created: UnorderedMap<String, u64>,
    pub updated: UnorderedMap<String, u64>,
}

Fields
status: UnorderedMap<String, Status>

Status is used to store the state of DID. There are two states of DID, valid and invalid.

contexts: UnorderedMap<String, Vec<String>>

context.

public_key: UnorderedMap<String, PublicKeyList>

It is used to store all public key information corresponding to this DID

authentication: UnorderedMap<String, Vec<u32>>controller: UnorderedMap<String, Vec<String>>

It is used to store all controller information corresponding to this DID. The controller has the authority to update the information of this did.

service: UnorderedMap<String, Vec<Service>>created: UnorderedMap<String, u64>

Used to store the creation time of DID

updated: UnorderedMap<String, u64>

Used to store the update time of DID

Methods

impl DID

pub fn reg_did_using_account(&mut self)

register DID

this method will store DID information on the chain

this method will output log information in the following format,

log information: "reg_did_using_account: did:near:abcde.testnet"

pub fn deactivate_did(&mut self)

deactivate_did DID

this method will update DID to deactive status, this means the DID is a invalid.

log information: "deactivate_did: did:near:abcde.testnet"

pub fn add_controller(&mut self, controller: String)

add_controller

this method will add a controller for the DID. controller is also a DID, it has the right to update the information of DID.

parameter controller is also a DID

log information: "add_controller, did: did:near:abcde.testnet, controller: did:near:abcdefg.testnet"

pub fn remove_controller(&mut self, controller: String)

remove_controller

this method will remove a controller for the DID. only the DID owner has the right to remove controller.

parameter controller is also a DID

log information: "remove_controller, did: did:near:abcde.testnet, controller: did:near:abcdefg.testnet"

pub fn add_key(&mut self, pk: Vec<u8>, controller: String)

add_key

this method will add a public key for the DID. only the DID owner has the right to add public key.

parameter controller is also a DID

log information: "add_key, did: did:near:abcde.testnet, public key: , controller: did:near:abcdefg.testnet"

pub fn deactivate_key(&mut self, pk: Vec<u8>)

deactivate_key

this method will update a public key to deactive status. only the DID owner has the right to invoke this method.

parameter pk is a public key

log information: "deactivate_key, did: did:near:abcde.testnet, public key: "

pub fn add_new_auth_key(&mut self, pk: Vec<u8>, controller: String)

add_new_auth_key

this method will add a public key to deactive status. only the DID owner has the right to invoke this method.

pk is a public key

controller is a DID

log information: "add_new_auth_key, did: did:near:abcde.testnet, public key: ,controller: did:near:abcdefg.testnet"

pub fn set_auth_key(&mut self, pk: Vec<u8>)

set_auth_key

this method will set the pk to authentication status. only the DID owner has the right to invoke this method.

pk is a public key

log information: "set_auth_key, did: did:near:abcde.testnet, public key: "

pub fn deactivate_auth_key(&mut self, pk: Vec<u8>)

deactivate_auth_key

this method will update the pk to authentication invalid status. only the DID owner has the right to invoke this method.

pk is a public key

log information: "deactivate_auth_key, did: did:near:abcde.testnet, public key: "

pub fn add_new_auth_key_by_controller(
&mut self,
did: String,
pk: Vec<u8>,
controller: String
)

add_new_auth_key_by_controller

this method will add new auth key by controller. need the controller's signature.

pk is controller's public key

log information: "add_new_auth_key_by_controller, did: did:near:abcde.testnet, public key: ,controller: did:near:abcdefg.testnet"

pub fn set_auth_key_by_controller(&mut self, did: String, pk: Vec<u8>)

set_auth_key_by_controller

this method will set the public key to auth key by controller. need the controller's signature.

pk is controller's public key

log information: "set_auth_key_by_controller, did: did:near:abcde.testnet, public key: "

pub fn deactivate_auth_key_by_controller(&mut self, did: String, pk: Vec<u8>)

deactivate_auth_key_by_controller

this method will update the auth public key to invalid status. need the controller's signature.

pk is controller's public key

log information: "deactivate_auth_key_by_controller, did: did:near:abcde.testnet, public key: "

pub fn add_service(
&mut self,
service_id: String,
service_type: String,
endpoint: String
)

add_service

this method will add service to the DID.

log information: "add_service, did: did:near:abcde.testnet, service id: 1234"

pub fn update_service(
&mut self,
service_id: String,
service_type: String,
endpoint: String
)

update_service

this method will update service.

log information: "update_service, did: did:near:abcde.testnet, service id: 1234"

pub fn remove_service(&mut self, service_id: String)

remove_service

this method will remove service.

log information: "remove_service, did: did:near:abcde.testnet, service id: 1234"

pub fn add_context(&mut self, context: Vec<String>)

add_context

this method will add context.

log information: "add_context, did: did:near:abcde.testnet, context: 1234"

pub fn remove_context(&mut self, context: Vec<String>)

remove_context

this method will remove context.

log information: "remove_context, did: did:near:abcde.testnet, context: 1234"

pub fn verify_signature(&self)

verify_signature

this method will verify a transaction is signed by did athentication key.

pub fn verify_controller(&self, did: String)

verify_controller

this method will verify a transaction is signed by did controller.

this method will verify a DID is or not the controller.

pub fn get_document(&self, did: String) -> Option<String>

get_document

this method query the did document, return json string.

this method query the DID information.